+++ src/components/common/CommonInfoPopup.vue
@@ -69,8 +69,8 @@
 </template>
 
 <script setup lang="ts">
-import CloseIcon from 'assets/icons/close.svg?component';
-import ArrowTopIcon from 'assets/icons/arrow-top.svg?component';
+import CloseIcon from 'assets/icons/basic/close.svg?component';
+import ArrowTopIcon from 'assets/icons/basic/arrow-top.svg?component';
 import type { PropType } from 'vue';
 
 interface InfoPopupSection {