+++ CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+# 0.1.1
+
+## Changes
+
+- Roadmap redesign
+- Added icons to header
+- 404/500 pages
+
+### Improved zoom experience
+
+- Overlays should stop opening when hovering on very busy map area
+- Overlays will stop opening when you're zooming or moving map
+
+## Technical features
+
+- You can now see release number in discord changelog
+- Removed generic favicon while we wait for proper one
+
 # 0.1.0
 
 This is initial release to test changelog generation and general CI/CD flow. Service is still pending Vatsim approval to announce proper testing.